Add untested vector-function class
vFunc is meant to represent a vector of functions in the sense of
the Extended Kalman Filter; each term in the vector manages the
transition of that variable. It allows for a more modular treatment
of the EKF implementation itself; rather than relying on the creation
of a singular, large, VectorXd -> VectorXd, you can just declare
and populate a vector like this.
Ideally, it'll be usable more or less like a standard vector in
the future -- if V is a function vector, A is a matrix, and x is
a vector, then A*V(x) would perform a valid operation. Should work.