Change goal to implementing Unscented Kalman Filter
Turns out it, or the Particle Filter or regular Kalman Filter,
are more suited to prettymuch any given application. As such,
we're switching to implementing that instead -- luckily, all
the work put in so far, focused on setting up the vector
function representation through vFunc, can be easily used in
any UKF implementation.