Add and Implement Private getter methods for GenSim
The only one that's an issue/non-obvious is GetCurrentState, which
is currently wrapped into the _measurement(~) call. I'm going to set
it up so that they're equivalent only if you pass two FALSE values
for the noise parameter options, giving an option to apply one or
another so should the user wish.